pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
Flask-Login = "^0.6.2"
Flask-SQLAlchemy = "^3.1.1"
relay = "^0.0.8"
Copreus = "^0.4.0"
bs4 = "^0.0.1"
bson = "^0.5.10"
pymongo = "^4.5.0"
discord = "^2.3.2"
django-user = "^0.2.1"
selenium = "^4.13.0"
python-dotenv = "^1.0.0"
"discord.py" = "^2.3.2"
openai = "^1.9.0"
passlib = "^1.7.4"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"